parameterized example
Each automaton in this family is a bottleneck of capacity n, facilitated by $O(n)$ states and actions.
The idea is to let all processes go to one of $n+1$ distinct nodes; afterwards, one has to play an action $aj$
identifying a node $j$ that is not occupied. This move wins, any other loses.

All of these models are **not** congrollable for arbitrary number of processes, but the $n$-fold product of the size-$m$ model is controllable iff $n\le m$.

The two python scripts generate (dot) NFAs and run commands on them,
respectively. See `examples/guess-empty-n/README.md` for details.